format data entry, screen cleaning. Version 1

+++ part.h
@@ -10,15 +10,15 @@ class Part                              // класс для хранения и
 {
 private:
     string partName;            // название детали
-    uint16_t article,           // артикул, соот-щий № детали в каталоге запчастей производителя
+    uint32_t article,           // артикул, соот-щий № детали в каталоге запчастей производителя
              availability;      // наличие на складе автомастерской
     float price;                // стоимость детали для покупки у автомастерской
 public:
-    Part(string p_partName, uint16_t p_article,                                 // конструктор
-         uint16_t p_availability, float p_price);
+    Part(string p_partName, uint32_t p_article,                                 // конструктор
+         uint32_t p_availability, float p_price);
 
-    void getInfoPart(string& p_partName, uint16_t& p_article,               // метод для возврата
-                       uint16_t& p_availability, float& p_price);           // значения полей
+    void getInfoPart(string& p_partName, uint32_t& p_article,               // метод для возврата
+                       uint32_t& p_availability, float& p_price);           // значения полей
     void edit();        //редактирование
 };
