+++ workorder.h
@@ -14,15 +14,15 @@ private:
            carModel,                // модель автомобиля
            orderInfo,               // информация о заказе
            status;                  // статус заказа
-    uint16_t orderNumber;   // номер заказа
+    uint32_t orderNumber;   // номер заказа
     float price;            // стоимость работ/заказа
 public:
-    WorkOrder(uint16_t p_orderNumber, string p_name,            // конструктор
+    WorkOrder(uint32_t p_orderNumber, string p_name,            // конструктор
               string p_carBrand, string p_carModel,
               string p_orderInfo, float p_price,
               string p_status);
 
-    void getInfoWorkOrder(uint16_t& p_orderNumber,              // метод для возврата
+    void getInfoWorkOrder(uint32_t& p_orderNumber,              // метод для возврата
                           string& p_name, string& p_carBrand,         // значения полей
                           string& p_carModel, string& p_orderInfo,
                           float& p_price,
